common commit of several modifications since 08/2014; intended to be used as starting point for migration to Git

+++ evaluate/sampleReader.h
@@ -11,7 +11,7 @@ private:
 
   std::stringstream channel0Filename, channel1Filename;
   std::ifstream timeFile, channel0File, channel1File;
-  double *channel0Buffer, *channel1Buffer;
+  float *channel0Buffer, *channel1Buffer;
   timespec bufferBegin, bufferEnd;
 
   timespec examinedSoFar;
